claude/pyproject.toml
[tool.poetry]
name = "claude"
version = "0.1.0"
description = ""
authors = ["kimkim <nkimura2021infotech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
